deduplicate filter function for issue cuplv/callback-verification#84
